app/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-extensions==3.2.1
django-filter==22.1
sqlparse==0.4.3
